services/loggingService/requirements.txt
python-dotenv
Flask
psycopg2-binary
pika
